ACCUMULO-3920 Convert more tests from mock
* Convert tests using mock to MiniAccumuloCluster
* Separate ITs from UTs, when possible
* Fix checkstyle/findbugs issues
