+++ test/src/main/java/org/apache/accumulo/test/replication/StatusMakerIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.accumulo.master.replication;
+package org.apache.accumulo.test.replication;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -28,10 +28,7 @@ import java.util.UUID;
 import org.apache.accumulo.core.client.BatchWriter;
 import org.apache.accumulo.core.client.BatchWriterConfig;
 import org.apache.accumulo.core.client.Connector;
-import org.apache.accumulo.core.client.Instance;
 import org.apache.accumulo.core.client.Scanner;
-import org.apache.accumulo.core.client.impl.Credentials;
-import org.apache.accumulo.core.client.security.tokens.PasswordToken;
 import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Mutation;
 import org.apache.accumulo.core.data.Value;
@@ -41,36 +38,35 @@ import org.apache.accumulo.core.replication.ReplicationSchema.OrderSection;
 import org.apache.accumulo.core.replication.ReplicationSchema.StatusSection;
 import org.apache.accumulo.core.replication.ReplicationTable;
 import org.apache.accumulo.core.security.Authorizations;
+import org.apache.accumulo.core.security.TablePermission;
+import org.apache.accumulo.master.replication.StatusMaker;
 import org.apache.accumulo.server.replication.StatusUtil;
 import org.apache.accumulo.server.replication.proto.Replication.Status;
 import org.apache.accumulo.server.util.ReplicationTableUtil;
+import org.apache.accumulo.test.functional.ConfigurableMacBase;
 import org.apache.hadoop.io.Text;
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TestName;
 
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 
-public class StatusMakerTest {
-
-  @Rule
-  public TestName test = new TestName();
+public class StatusMakerIT extends ConfigurableMacBase {
 
   private Connector conn;
 
   @Before
   public void setupInstance() throws Exception {
-    Instance inst = new org.apache.accumulo.core.client.mock.MockInstance(test.getMethodName());
-    Credentials creds = new Credentials("root", new PasswordToken(""));
-    conn = inst.getConnector(creds.getPrincipal(), creds.getToken());
+    conn = getConnector();
+    ReplicationTable.setOnline(conn);
+    conn.securityOperations().grantTablePermission(conn.whoami(), ReplicationTable.NAME, TablePermission.WRITE);
+    conn.securityOperations().grantTablePermission(conn.whoami(), ReplicationTable.NAME, TablePermission.READ);
   }
 
   @Test
   public void statusRecordsCreated() throws Exception {
-    String sourceTable = "source";
+    String sourceTable = testName.getMethodName();
     conn.tableOperations().create(sourceTable);
     ReplicationTableUtil.configureMetadataTable(conn, sourceTable);
 
@@ -117,7 +113,7 @@ public class StatusMakerTest {
 
   @Test
   public void openMessagesAreNotDeleted() throws Exception {
-    String sourceTable = "source";
+    String sourceTable = testName.getMethodName();
     conn.tableOperations().create(sourceTable);
     ReplicationTableUtil.configureMetadataTable(conn, sourceTable);
 
@@ -153,7 +149,7 @@ public class StatusMakerTest {
 
   @Test
   public void closedMessagesAreDeleted() throws Exception {
-    String sourceTable = "source";
+    String sourceTable = testName.getMethodName();
     conn.tableOperations().create(sourceTable);
     ReplicationTableUtil.configureMetadataTable(conn, sourceTable);
 
@@ -196,7 +192,7 @@ public class StatusMakerTest {
 
   @Test
   public void closedMessagesCreateOrderRecords() throws Exception {
-    String sourceTable = "source";
+    String sourceTable = testName.getMethodName();
     conn.tableOperations().create(sourceTable);
     ReplicationTableUtil.configureMetadataTable(conn, sourceTable);
